backend/requirements.txt
Django~=3.1.4
djangorestframework~=3.12.4
pathlib~=1.0.1
django-cors-headers
Pillow
django-environ
